bmath

+++ include/bgen/meta.h
@@ -55,8 +55,8 @@ struct bgen_vm
     struct bgen_str const*  id;         /**< Variant identification. */
     struct bgen_str const*  rsid;       /**< RSID. */
     struct bgen_str const*  chrom;      /**< Chromossome name. */
-    unsigned                position;   /**< Base-pair position. */
-    unsigned                nalleles;   /**< Number of alleles. */
+    uint32_t                position;   /**< Base-pair position. */
+    uint16_t                nalleles;   /**< Number of alleles. */
     struct bgen_str const** allele_ids; /**< Allele ids. */
 };
